mdx loader refactoring
